requirements.txt
asn1crypto==0.24.0
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
future==0.17.1
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.13.1
idna==2.8
oauthlib==3.0.2
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
python-telegram-bot==11.1.0
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.3
